Add findRoom function to go to a specific room
