+++ treasure-hunt/src/util/changeName.js
@@ -3,9 +3,9 @@
 
 import axios from "axios";
 import { sleep } from "./sleep";
+import { findRoom } from "./findRoom";
 
 const graph = require("../data/graph.json");
-const reverse_dir = { n: "s", s: "n", e: "w", w: "e" };
 const BASE_URL = "https://lambda-treasure-hunt.herokuapp.com/api/adv/";
 const INIT_URL = BASE_URL + "init/";
 const MOVE_URL = BASE_URL + "move/";
@@ -20,7 +20,8 @@ export const changeName = async () => {
     cooldown = init.data.cooldown;
     await sleep(cooldown * 1000);
     let current_room = init.data.room_id;
-    const path = findPirate(current_room);
+    // const path = findPirate(current_room);
+    const path = findRoom(current_room, 467);
     console.log(path);
     for (let i = 0; i < path.length; i++) {
       const exits = graph[current_room]["directions"];
@@ -48,34 +49,4 @@ export const changeName = async () => {
   } catch (error) {
     console.error(error);
   }
-};
-
-const findPirate = room => {
-  // Use DFS to get to room 467
-  const queue = [];
-  const visited_rooms = new Set();
-
-  queue.push([[null, room]]);
-  while (queue.length > 0) {
-    const path = queue.shift();
-    const room = path[path.length - 1][1];
-    if (room === 467) {
-      const traversal_path = [];
-      for (let i = 0; i < path.length; i++) {
-        const pair = path[i];
-        if (pair[0] !== null) traversal_path.push(pair[0]);
-      }
-      return traversal_path;
-    }
-    if (!visited_rooms.has(room)) {
-      visited_rooms.add(room);
-      const exits = graph[room]["directions"];
-      for (const direction in exits) {
-        const current_path = [...path];
-        current_path.push([direction, exits[direction]]);
-        queue.push(current_path);
-      }
-    }
-  }
-  return [];
 };
